+++ resources/js/components/GraphqlMutation.vue
@@ -78,6 +78,7 @@ export default {
         initialVariables: {},
         data: {},
         mutate: () => null,
+        redirectUrl: '',
     }),
 
     render() {
@@ -99,6 +100,7 @@ export default {
     created() {
         this.initialVariables = JSON.parse(JSON.stringify(this.variables))
         this.data = this.variables
+        this.redirectUrl = this.redirect
 
         if (this.debounce) {
             this.mutate = useDebounceFn(async () => await this.mutateFn(), this.debounce)
@@ -185,7 +187,7 @@ export default {
                 })
 
                 if (response.errors) {
-                    return
+                    return false
                 }
 
                 if (this.callback) {
@@ -202,11 +204,11 @@ export default {
                     self.mutated = false
                 }, 2500)
 
-                if (!this.redirect && this.notify.message) {
+                if (!this.redirectUrl && this.notify.message) {
                     Notify(this.notify.message, this.notify.type ?? 'success')
                 }
 
-                if (this.redirect) {
+                if (this.redirectUrl) {
                     if (this.notify.message) {
                         document.addEventListener(
                             'vue:loaded',
@@ -216,8 +218,10 @@ export default {
                             { once: true },
                         )
                     }
-                    Turbo.visit(window.url(this.redirect))
+                    Turbo.visit(window.url(this.redirectUrl))
                 }
+
+                return response
             } catch (error) {
                 console.error(error)
                 this.error = error.message