+++ resources/js/fetch.js
@@ -42,7 +42,7 @@ export const rapidezAPI = (window.rapidezAPI = async (method, endpoint, data = {
                 Store: window.config.store_code,
                 Authorization: token.value ? `Bearer ${token.value}` : null,
                 'Content-Type': 'application/json',
-                'X-CSRF-Token': window.app.csrfToken,
+                'X-CSRF-Token': window.app.config.globalProperties.csrfToken,
             },
             options?.headers || {},
         ),