+++ resources/views/listing/partials/filter/price.blade.php
@@ -1,6 +1,6 @@
 <ais-range-input
     v-if="listingSlotProps.rangeAttributes.includes(filter.code)"
-    :attribute="filter.code == 'price' ? 'prices.'+($root?.user?.group_id || 0)+'.min_price' : filter.code"
+    :attribute="filter.code == 'price' ? 'prices.'+(user.value?.group_id || 0)+'.min_price' : filter.code"
     class="relative -mx-1 -mt-1"
 >
     <template v-slot="{ currentRefinement, range, canRefine, refine, sendEvent }">